Formatted the logged in client logic in Store Controller, I did nothing in the second half.
All but 2 things are the same I just moved some things around:
I went back to the filter for the cart ove the query because I wasn't able to get it to work, if you want to keep that you might have more luck than me.
I moved some of the logic into a method which I placed at the bottom of the page.

I added some logic to the Order Controller where Basket Items are either removed from the cart or deleted, and the basket is deleted.

I still have to do more testing, but on my first couple passes everything was working.